# contracts/coding-labs/src/models/LabRunnerConfigDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LabRunnerConfigDto = {
    timeoutMs: number;
    memoryMb?: number;
    entryFnName?: string;
    nodeVersion?: string;
};