# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/labs/{labId}": {
      "get": {
        "operationId": "LabsController_findOne",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      },
      "patch": {
        "operationId": "LabsController_update",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      },
      "delete": {
        "operationId": "LabsController_remove",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      }
    },
    "/labs/{labId}/versions/draft": {
      "post": {
        "operationId": "LabVersionsController_createDraft",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabVersionMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Versions"
        ]
      }
    },
    "/labs/{labId}/versions": {
      "get": {
        "operationId": "LabVersionsController_findAll",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HandsOnLabVersionMongo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Versions"
        ]
      }
    },
    "/labs/{labId}/versions/{versionId}": {
      "get": {
        "operationId": "LabVersionsController_findOne",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabVersionMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Versions"
        ]
      },
      "patch": {
        "operationId": "LabVersionsController_patchDraft",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabVersionMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Versions"
        ]
      }
    },
    "/labs/{labId}/versions/{versionId}/verify": {
      "post": {
        "operationId": "LabVersionsController_verify",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationDto"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Versions"
        ]
      }
    },
    "/labs/{labId}/versions/{versionId}/publish": {
      "post": {
        "operationId": "LabVersionsController_publish",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabVersionMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Versions"
        ]
      }
    },
    "/embeds": {
      "post": {
        "operationId": "LabEmbedsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmbedDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabEmbedMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Embeds"
        ]
      },
      "get": {
        "operationId": "LabEmbedsController_findAll",
        "parameters": [
          {
            "name": "labId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workshopId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workshopDocumentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HandsOnLabEmbedMongo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Embeds"
        ]
      }
    },
    "/embeds/{embedId}": {
      "delete": {
        "operationId": "LabEmbedsController_remove",
        "parameters": [
          {
            "name": "embedId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lab Embeds"
        ]
      }
    }
  },
  "info": {
    "title": "NGX coding-labs API",
    "description": "",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ComparatorDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "deepEqual",
              "strictEqual",
              "numberTolerance",
              "stringNormalized",
              "custom"
            ]
          },
          "tolerance": {
            "type": "number"
          },
          "normalizeWhitespace": {
            "type": "boolean"
          },
          "ignoreCase": {
            "type": "boolean"
          },
          "customComparatorId": {
            "type": "string"
          }
        },
        "required": [
          "kind"
        ]
      },
      "LabTestCaseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "io",
              "unit"
            ]
          },
          "input": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "expected": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "comparator": {
            "$ref": "#/components/schemas/ComparatorDto"
          },
          "testCode": {
            "type": "string"
          },
          "framework": {
            "type": "string",
            "enum": [
              "jest",
              "vitest"
            ]
          }
        },
        "required": [
          "name",
          "kind"
        ]
      },
...
      "HandsOnLabMongo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "workshopId": {
            "type": "string"
          },
          "workshopDocumentGroupId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "intro",
              "easy",
              "medium",
              "hard"
            ]
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "currentDraftVersionId": {
            "type": "string"
          },
          "latestPublishedVersionId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "archivedAt": {
            "type": "string"
          },
          "archivedBy": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "workshopId",
          "slug",
          "title",
          "tags",
          "status",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy"
        ]
      },
      "UpdateLabDto": {
        "type": "object",
        "properties": {
          "workshopDocumentGroupId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "tags": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "intro",
              "easy",
              "medium",
              "hard"
            ]
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "ReferenceSolutionDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "notesMarkdown": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "LabRunnerConfigDto": {
        "type": "object",
        "properties": {
          "timeoutMs": {
            "type": "number"
          },
          "memoryMb": {
            "type": "number"
          },
          "entryFnName": {
            "type": "string"
          },
          "nodeVersion": {
            "type": "string"
          }
        },
        "required": [
          "timeoutMs"
        ]
      },
      "CreateDraftVersionDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "typescript",
              "javascript"
            ]
          },
          "promptMarkdown": {
            "type": "string"
          },
          "hints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "starterCode": {
            "type": "string"
          },
          "referenceSolution": {
            "$ref": "#/components/schemas/ReferenceSolutionDto"
          },
          "sampleTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabTestCaseDto"
            }
          },
          "hiddenTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabTestCaseDto"
            }
          },
          "runner": {
            "$ref": "#/components/schemas/LabRunnerConfigDto"
          },
          "contentHash": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          }
        },
        "required": [
          "createdBy"
        ]
      },
      "HandsOnLabVersionMongo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "labId": {
            "type": "string"
          },
          "versionNumber": {
            "type": "number"
          },
          "isDraft": {
            "type": "boolean"
          },
          "language": {
            "type": "string",
            "enum": [
              "typescript",
              "javascript"
            ]
          },
          "promptMarkdown": {
            "type": "string"
          },
          "hints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "starterCode": {
            "type": "string"
          },
          "referenceSolution": {
            "$ref": "#/components/schemas/ReferenceSolutionDto"
          },
          "sampleTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabTestCaseDto"
            }
          },
          "hiddenTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabTestCaseDto"
            }
          },
          "runner": {
            "$ref": "#/components/schemas/LabRunnerConfigDto"
          },
          "publishedAt": {
            "type": "string"
          },
          "publishedBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "labId",
          "versionNumber",
          "isDraft",
          "language",
          "promptMarkdown",
          "starterCode",
          "sampleTests",
          "hiddenTests",
          "runner",
          "createdAt",
          "createdBy"
        ]
      },
      "UpdateDraftVersionDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "typescript",
              "javascript"
            ]
          },
          "promptMarkdown": {
            "type": "string"
          },
          "hints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "starterCode": {
            "type": "string"
          },
          "referenceSolution": {
            "$ref": "#/components/schemas/ReferenceSolutionDto"
          },
          "sampleTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabTestCaseDto"
            }
          },
          "hiddenTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabTestCaseDto"
            }
          },
          "runner": {
            "$ref": "#/components/schemas/LabRunnerConfigDto"
          },
          "contentHash": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "expectedContentHash": {
            "type": "string",
            "description": "Hash from the loaded draft; prevents stale overwrites"
          }
        },
        "required": [
          "expectedContentHash"
        ]
      },
      "TestResultDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "suite": {
            "type": "string",
            "enum": [
              "sample",
              "hidden"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "passed",
              "failed",
              "error",
              "timeout"
            ]
          },
          "durationMs": {
            "type": "number"
          },
          "actual": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "expected": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "suite",
          "status",
          "durationMs"
        ]
      },
      "VerificationDto": {
        "type": "object",
        "properties": {
          "passed": {
            "type": "boolean"
          },
          "contentHash": {
            "type": "string"
          },
          "totalTests": {
            "type": "number"
          },
          "passedTests": {
            "type": "number"
          },
          "durationMs": {
            "type": "number"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestResultDto"
            }
          }
        },
        "required": [
          "passed",
          "contentHash",
          "totalTests",
          "passedTests",
          "durationMs",
          "results"
        ]
      },
      "PublishVersionDto": {
        "type": "object",
        "properties": {
          "expectedContentHash": {
            "type": "string"
          },
          "publishedBy": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "expectedContentHash",
          "publishedBy"
        ]
      },
      "CreateEmbedDto": {
        "type": "object",
        "properties": {
          "labId": {
            "type": "string"
          },
          "workshopId": {
            "type": "string"
          },
          "workshopDocumentId": {
            "type": "string"
          },
          "blockId": {
            "type": "string"
          },
          "blockType": {
            "type": "string",
            "enum": [
              "handsOnLab"
            ],
            "default": "handsOnLab"
          },
          "pinnedVersionId": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          }
        },
        "required": [
          "labId",
          "workshopId",
          "workshopDocumentId",
          "blockId",
          "createdBy"
        ]
      },
      "HandsOnLabEmbedMongo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "labId": {
            "type": "string"
          },
          "workshopId": {
            "type": "string"
          },
          "workshopDocumentId": {
            "type": "string"
          },
          "blockId": {
            "type": "string"
          },
          "blockType": {
            "type": "string"
          },
          "pinnedVersionId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "labId",
          "workshopId",
          "workshopDocumentId",
          "blockId",
          "blockType",
          "createdAt",
          "createdBy"
        ]
      }
    }
  }
}